update this very script’s status, too
This script now has its own stoplight widget on the dashboard. If an exception is raised while trying to get data and send it to the dashboard, the script tries to tell the stoplight widget that it failed.

Why? It would be bad to have old information on the dashboard and be unaware of it.