preliminary Freckle hours pushing to dashboard
The data is successfully pushed to a newly-created line graph, but subsequent pushes either do nothing or draw lines in the wrong order, rather than updating the existing lines.

The points of the times are also displayed a little to the left, because the Date turned into Unix time is four hours before midnight – I think that’s because of our timezone.

The data points are reversed so they are drawn in chronological order.

Commented out updating build statuses to make testing the freckle hours easier.