removed seconds from time in “Updating” status message
The dashboard itself doesn’t show seconds, so there is nothing to compare the displayed seconds to.

What’s more, the seconds part comes from the time on the computer running the update script, not necessarily the time shown on the dashboard. (And the time shown on the dashboard appears to not be the same as the time on the computer showing the dashboard web page – that, or there is a delay in the time updating.)