abstracted spotlight widget colors from numeric values
This makes it easier to read the list of build_status_ratings, and to a lesser extent, the status_ratings in LeftronicUpdateStatusUpdater.

I didn’t put this in the actual leftronic client library as #push_spotlight because that library has to be backwards-compatible, so I would have to be too careful that my method would be future-proof. Though perhaps a monkey-patch or wrapper of the library would be cleaner than this – it’s a little weird having a global constant (STOPLIGHT_COLOR_NUMBERS) that’s only related to the Leftronic dashboard.